Moved functional tests to another package
