+++ src/test/java/fi/metatavu/metaform/test/functional/FileUploadResponse.java
@@ -1,4 +1,4 @@
-package fi.metatavu.metaform.server;
+package fi.metatavu.metaform.test.functional;
 
 import java.util.UUID;
